Issue #206: change score type to float

+++ quarkus/datasource/src/main/java/io/openk9/datasource/searcher/SearcherService.java
@@ -624,7 +624,7 @@ public class SearcherService extends BaseSearchService implements Searcher {
 			}
 
 			if (map.containsKey("score")) {
-				builder.setScore((Double)map.get("score"));
+				builder.setScore((Float)map.get("score"));
 			}
 
 			if (map.containsKey("keywordKey")) {
